caroptionshop-back-django/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
psycopg==3.2.3
sqlparse==0.5.2
typing_extensions==4.12.2
